Moving DocumentScreen from into screens
Moving from `components/DocumentScreen` to `screens/DocumentScreen`.

Will follow up with viewer refactoring.
