+++ ui/src/components/Preview/Preview.jsx
@@ -10,8 +10,8 @@ import getPath from 'src/util/getPath';
 import { fetchCollection, fetchEntity, fetchDocument } from 'src/actions';
 import CollectionInfo from 'src/components/CollectionScreen/CollectionInfo';
 import EntityInfo from 'src/screens/EntityScreen/EntityInfo';
-import DocumentInfo from 'src/components/DocumentScreen/DocumentInfo';
-import DocumentContent from 'src/components/DocumentScreen/DocumentContent';
+import DocumentInfo from 'src/screens/DocumentScreen/DocumentInfo';
+import DocumentContent from 'src/screens/DocumentScreen/DocumentContent';
 import SectionLoading from 'src/components/common/SectionLoading';
 
 import './Preview.css';
