+++ ui/src/screens/DocumentScreen/DocumentContent.scss
