+++ ui/src/screens/DocumentScreen/DocumentRedirectScreen.jsx
