+++ ui/src/screens/DocumentScreen/viewers/EmailViewer.scss
