+++ ui/src/screens/DocumentScreen/viewers/FolderViewer.scss
