+++ ui/src/screens/DocumentScreen/viewers/HtmlViewer.scss
