+++ ui/src/screens/DocumentScreen/viewers/ImageViewer.scss
