+++ ui/src/screens/DocumentScreen/viewers/PdfViewer.scss
