+++ ui/src/screens/DocumentScreen/viewers/TextViewer.scss
